rename textarea to input and use that in modal; remove other input component
