+++ src/components/experiment-tracking/run-details-modal/run-details-modal.scss
@@ -1,3 +1,7 @@
+.pipeline-settings-modal--experiment-tracking .pipeline-settings-modal__name {
+  margin-top: 0;
+}
+
 .run-details-modal-button-wrapper {
   align-items: baseline;
   display: flex;
