+++ src/components/ui/input/input.scss
@@ -3,11 +3,32 @@
 .input {
   @extend %input;
 
+  min-height: auto;
+  resize: none;
+
   &--large {
     @extend %input--large;
+
+    height: 24px;
   }
 
   &--small {
     @extend %input--small;
+
+    height: 20px;
   }
+}
+
+.input-character-count {
+  align-items: baseline;
+  color: var(--color-text-faded);
+  display: flex;
+  font-size: 16px;
+  justify-content: flex-end;
+  margin-bottom: 6px;
+  width: 100%;
+}
+
+.input-number-characters {
+  color: var(--color-text);
 }
