--- src/components/ui/textarea/textarea.scss
@@ -1,34 +1,0 @@
-@import '../../../styles/extends';
-
-.textarea {
-  @extend %input;
-
-  min-height: auto;
-  resize: none;
-
-  &--large {
-    @extend %input--large;
-
-    height: 24px;
-  }
-
-  &--small {
-    @extend %input--small;
-
-    height: 20px;
-  }
-}
-
-.textarea-character-count {
-  align-items: baseline;
-  color: var(--color-text-faded);
-  display: flex;
-  font-size: 16px;
-  justify-content: flex-end;
-  margin-bottom: 6px;
-  width: 100%;
-}
-
-.textarea-number-characters {
-  color: var(--color-text);
-}
